scripts/requirements.txt
-i https://pypi.org/simple
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
click==7.0
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
idna==2.8
protobuf==3.10.0
pyasn1-modules==0.2.7
pyasn1==0.4.7
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.6
